Fix tools build script.
